Add simple dashboard
This commit initiate a dashboard for our source of data. It uses the new
WebClient to fetch the data and aggregate power consumption per zone. It
saves this history in a MongoDB collection.

The dashboard also displays the current power consumption per zone but
doesn't provide live update at the moment.