Advanced SSE use case with Thymeleaf
This commit demonstrates a feature of Thymeleaf that allows to let
thymeleaf subscribe on a hot source of data. We reuse the data coming
from the backend to extract only the items where the power consumption
is 0.

We can simulate a power failure by changing how sample data are
generated. For instance via an HTTP call:

http :8081/application/zones/one powerLow=0 powerHigh=0